+++ packages/vtable/examples/menu.ts
@@ -77,6 +77,10 @@ export const menus = [
       },
       {
         path: 'list',
+        name: 'list-empty-tip'
+      },
+      {
+        path: 'list',
         name: 'list-height-adaptive'
       },
       {
@@ -257,6 +261,10 @@ export const menus = [
       {
         path: 'pivot-chart',
         name: 'pivotChart-legend'
+      },
+      {
+        path: 'pivot-chart',
+        name: 'pivotChart-two-legends'
       },
       {
         path: 'pivot-chart',