+++ packages/vtable/src/core/BaseTable.ts
@@ -127,6 +127,7 @@ import { CustomCellStylePlugin, mergeStyle } from '../plugins/custom-cell-style'
 import { hideCellSelectBorder, restoreCellSelectBorder } from '../scenegraph/select/update-select-border';
 import type { ITextGraphicAttribute } from '@src/vrender';
 import type { ISortedMapItem } from '../data/DataSource';
+import { hasAutoImageColumn } from '../layout/layout-helper';
 
 const { toBoxArray } = utilStyle;
 const { isTouchEvent } = event;
@@ -195,6 +196,8 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
   customCellStylePlugin: CustomCellStylePlugin;
 
   columnWidthComputeMode?: 'normal' | 'only-header' | 'only-body';
+
+  _hasAutoImageColumn?: boolean;
 
   constructor(container: HTMLElement, options: BaseTableConstructorOptions = {}) {
     super();
@@ -317,7 +320,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
     internalProps.columnResizeMode = columnResizeMode;
     internalProps.rowResizeMode = rowResizeMode;
-    internalProps.dragHeaderMode = dragHeaderMode;
+    internalProps.dragHeaderMode = dragHeaderMode ?? 'none';
     internalProps.renderChartAsync = renderChartAsync;
     setBatchRenderChartCount(renderChartAsyncBatchCount);
     internalProps.overscrollBehavior = overscrollBehavior ?? 'auto';
@@ -395,11 +398,22 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     this.eventManager = new EventManager(this);
 
     if (options.legends) {
-      internalProps.legends = createLegend(options.legends, this);
-      this.scenegraph.tableGroup.setAttributes({
-        x: this.tableX,
-        y: this.tableY
-      });
+      internalProps.legends = [];
+      if (Array.isArray(options.legends)) {
+        for (let i = 0; i < options.legends.length; i++) {
+          internalProps.legends.push(createLegend(options.legends[i], this));
+        }
+        this.scenegraph.tableGroup.setAttributes({
+          x: this.tableX,
+          y: this.tableY
+        });
+      } else {
+        internalProps.legends.push(createLegend(options.legends, this));
+        this.scenegraph.tableGroup.setAttributes({
+          x: this.tableX,
+          y: this.tableY
+        });
+      }
     }
 
     //原有的toolTip提示框处理，主要在文字绘制不全的时候 出来全文本提示信息 需要加个字段设置是否有效
@@ -469,11 +483,14 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
   resize() {
     this._updateSize();
-    if (this.internalProps.legends) {
-      this.internalProps.legends.resize();
-    }
+    this.internalProps.legends?.forEach(legend => {
+      legend?.resize();
+    });
     if (this.internalProps.title) {
       this.internalProps.title.resize();
+    }
+    if (this.internalProps.emptyTip) {
+      this.internalProps.emptyTip.resize();
     }
     // this.stateManager.checkFrozen();
     this.scenegraph.resize();
@@ -1102,6 +1119,12 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
       this._clearRowRangeHeightsMap(row);
     }
   }
+
+  setRowHeight(row: number, height: number) {
+    this.scenegraph.setRowHeight(row, height);
+    this.internalProps._heightResizedRowMap.add(row); // add resize tag
+  }
+
   /**
    * 获取指定行范围的总高度
    * @param startCol
@@ -1290,6 +1313,11 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     if (!skipCheckFrozen) {
       this.stateManager.checkFrozen();
     }
+  }
+
+  setColWidth(col: number, width: number) {
+    this.scenegraph.setColWidth(col, width);
+    this.internalProps._widthResizedColMap.add(col); // add resize tag
   }
 
   /**
@@ -2027,8 +2055,13 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     internalProps.handler?.release?.();
     // internalProps.scrollable?.release?.();
     internalProps.focusControl?.release?.();
-    internalProps.legends?.release();
+    internalProps.legends?.forEach(legend => {
+      legend?.release();
+    });
     internalProps.title?.release();
+    internalProps.title = null;
+    internalProps.emptyTip?.release();
+    internalProps.emptyTip = null;
     internalProps.layoutMap.release();
     if (internalProps.releaseList) {
       internalProps.releaseList.forEach(releaseObj => releaseObj?.release?.());
@@ -2059,6 +2092,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
    */
   updateOption(options: BaseTableConstructorOptions) {
     (this.options as BaseTable['options']) = options;
+    this._hasAutoImageColumn = undefined;
     const {
       // rowCount = 0,
       // colCount = 0,
@@ -2150,7 +2184,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
     internalProps.columnResizeMode = columnResizeMode;
     internalProps.rowResizeMode = rowResizeMode;
-    internalProps.dragHeaderMode = dragHeaderMode;
+    internalProps.dragHeaderMode = dragHeaderMode ?? 'none';
     internalProps.renderChartAsync = renderChartAsync;
     setBatchRenderChartCount(renderChartAsyncBatchCount);
     internalProps.overscrollBehavior = overscrollBehavior ?? 'auto';
@@ -2197,8 +2231,13 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
         : 10;
     // 生成scenegraph
     this._vDataSet = new DataSet();
-    internalProps.legends?.release();
+    internalProps.legends?.forEach(legend => {
+      legend?.release();
+    });
     internalProps.title?.release();
+    internalProps.title = null;
+    internalProps.emptyTip?.release();
+    internalProps.emptyTip = null;
     internalProps.layoutMap.release();
     this.scenegraph.clearCells();
     this.scenegraph.updateComponent();
@@ -2209,11 +2248,22 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     // this.eventManager = new EventManager(this);
     this.eventManager.updateEventBinder();
     if (options.legends) {
-      internalProps.legends = createLegend(options.legends, this);
-      this.scenegraph.tableGroup.setAttributes({
-        x: this.tableX,
-        y: this.tableY
-      });
+      internalProps.legends = [];
+      if (Array.isArray(options.legends)) {
+        for (let i = 0; i < options.legends.length; i++) {
+          internalProps.legends.push(createLegend(options.legends[i], this));
+        }
+        this.scenegraph.tableGroup.setAttributes({
+          x: this.tableX,
+          y: this.tableY
+        });
+      } else {
+        internalProps.legends.push(createLegend(options.legends, this));
+        this.scenegraph.tableGroup.setAttributes({
+          x: this.tableX,
+          y: this.tableY
+        });
+      }
     }
     // if (options.title) {
     //   internalProps.title = new Title(options.title, this);
@@ -2666,11 +2716,27 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     const { scrollLeft, scrollTop } = this;
     cellRanges.forEach((cellRange: CellRange, index: number) => {
       if (cellRange.start.col === cellRange.end.col && cellRange.start.row === cellRange.end.row) {
-        this.stateManager.updateSelectPos(cellRange.start.col, cellRange.start.row, false, index >= 1);
+        this.stateManager.updateSelectPos(
+          cellRange.start.col,
+          cellRange.start.row,
+          false,
+          index >= 1,
+          false,
+          false,
+          true
+        );
       } else {
-        this.stateManager.updateSelectPos(cellRange.start.col, cellRange.start.row, false, index >= 1);
+        this.stateManager.updateSelectPos(
+          cellRange.start.col,
+          cellRange.start.row,
+          false,
+          index >= 1,
+          false,
+          false,
+          true
+        );
         this.stateManager.updateInteractionState(InteractionState.grabing);
-        this.stateManager.updateSelectPos(cellRange.end.col, cellRange.end.row, false, index >= 1);
+        this.stateManager.updateSelectPos(cellRange.end.col, cellRange.end.row, false, index >= 1, false, false, true);
       }
       this.stateManager.endSelectCells(false);
       this.stateManager.updateInteractionState(InteractionState.default);
@@ -3514,22 +3580,29 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
    * @returns
    */
   _canDragHeaderPosition(col: number, row: number): boolean {
-    if (this.isHeader(col, row) && this.stateManager.isSelected(col, row)) {
+    if (
+      this.isHeader(col, row) &&
+      (this.stateManager.isSelected(col, row) ||
+        this.options.select?.disableHeaderSelect ||
+        this.options.select?.disableSelect)
+    ) {
       if (this.internalProps.frozenColDragHeaderMode === 'disabled' && this.isFrozenColumn(col)) {
         return false;
       }
-      const selectRange = this.stateManager.select.ranges[0];
-      //判断是否整行或者整列选中
-      if (this.isColumnHeader(col, row)) {
-        if (selectRange.end.row !== this.rowCount - 1) {
+      if (this.stateManager.isSelected(col, row)) {
+        const selectRange = this.stateManager.select.ranges[0];
+        //判断是否整行或者整列选中
+        if (this.isColumnHeader(col, row)) {
+          if (selectRange.end.row !== this.rowCount - 1) {
+            return false;
+          }
+        } else if (this.isRowHeader(col, row)) {
+          if (selectRange.end.col !== this.colCount - 1) {
+            return false;
+          }
+        } else {
           return false;
         }
-      } else if (this.isRowHeader(col, row)) {
-        if (selectRange.end.col !== this.colCount - 1) {
-          return false;
-        }
-      } else {
-        return false;
       }
       const define = this.getHeaderDefine(col, row);
       if (!define) {
@@ -3966,28 +4039,12 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
   }
 
   hasAutoImageColumn() {
-    const bodyColumn = (this.internalProps.layoutMap.columnObjects as ColumnData[]).find((column: ColumnData) => {
-      if (
-        (column.cellType === 'image' || column.cellType === 'video' || typeof column.cellType === 'function') &&
-        (column.define as ImageColumnDefine).imageAutoSizing
-      ) {
-        return true;
-      }
-      return false;
-    });
-    const headerObj = (this.internalProps.layoutMap.headerObjects as HeaderData[]).find((column: HeaderData) => {
-      if (column) {
-        if (
-          (column.headerType === 'image' || column.headerType === 'video' || typeof column.headerType === 'function') &&
-          (column.define as ImageColumnDefine).imageAutoSizing
-        ) {
-          return true;
-        }
-      }
-      return false;
-    });
-    return bodyColumn || headerObj;
+    if (this._hasAutoImageColumn === undefined) {
+      this._hasAutoImageColumn = hasAutoImageColumn(this);
+    }
+    return this._hasAutoImageColumn;
   }
+
   /** 获取当前hover单元格的图表实例。这个方法hover实时获取有点缺陷：鼠标hover到单元格上触发了 chart.ts中的activate方法 但此时this.stateManager.hover?.cellPos?.col还是-1 */
   _getActiveChartInstance() {
     // 根据hover的单元格位置 获取单元格实例 拿到chart图元