+++ packages/vtable/CHANGELOG.md
@@ -1,6 +1,61 @@
 # Change Log - @visactor/vtable
 
-This log was last generated on Tue, 21 May 2024 12:41:50 GMT and should not be manually modified.
+This log was last generated on Tue, 28 May 2024 12:24:56 GMT and should not be manually modified.
+
+## 1.1.0
+Tue, 28 May 2024 12:24:56 GMT
+
+### Minor changes
+
+- feat: add empty tip #1782
+
+
+
+### Updates
+
+- fix: add update delete record api should maintain beforeChangedRecordsMap #1780
+
+
+- fix: when set disableSelect table should support drag header #1800
+
+
+- fix: fix tree create problem in getDataCellPath()
+- fix: fix left axis index in horizontal pivot chart
+
+## 1.0.3
+Fri, 24 May 2024 10:31:16 GMT
+
+### Updates
+
+- refactor: legends support ser array form #1740
+
+
+- fix: first col tree mode compute col width error #1778
+
+
+
+## 1.0.2
+Fri, 24 May 2024 10:08:07 GMT
+
+### Updates
+
+- fix: use table option in hasAutoImageColumn()
+- fix: axis size align with vrender-component #1784
+- fix: fix lineClamp config in computeRowsHeight() #1772
+- fix: fix progress cell create problem in vtable-export #1787
+- fix: ignore cell merge in selectCells()
+- feat: add setRowHeight&setColWidth api
+
+## 1.0.1
+Wed, 22 May 2024 12:21:05 GMT
+
+### Updates
+
+- feat: tree mode  can set icon #1697
+
+
+- fix: ignore cell merge in selectCells()
+- feat: add setRowHeight&setColWidth api
 
 ## 1.0.0
 Tue, 21 May 2024 12:41:50 GMT